% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialEst.R
\name{MainEff}
\alias{MainEff}
\title{main effects estimate  in psiLearn}
\usage{
MainEff(X,A,p)
}
\arguments{
\item{X}{n*p input matrix.}

\item{A}{a vector of n entries coded 1 and -1 for the treatment assignments.}

\item{R}{a vector of outcome variable, larger is more desirable.}
}
\value{
It returns  the main effects estimated using weighted linear regression.
 \item{fitval}{the estimated main effects}
}
\description{
Suppose the main effects is linear, then return the fitted values in the weighted linear regression which
             is the main effects. Then the residual will be the interaction effects between the treatment and the "reward" plus the
             zero mean random effects.
}
\author{
MingyangLiu <liux3941@umn.edu>
}
