% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.R
\name{pre.psi}
\alias{pre.psi}
\title{predict treatment in psi-Learning}
\usage{
pre.psi(fitobj,Xnew,X=NULL,kernel='linear')
}
\arguments{
\item{fitobj}{model of class from psi-Learning}

\item{Xnew}{new data for prediction, mainly \eqn{n} by \eqn{p} input matrix.}

\item{X}{in linear case this is NULL and in 'rbf' kernel case this is the training data}

\item{kernel}{kernel function for psi-Learning, can be \code{'linear'} or \code{'rbf'} (radial basis kernel), default is \code{'linear'}.When using \code{'rbf'} , the bandwidth parameter
sigma is asked to provide}
}
\value{
predicted treatment for each new observation
}
\description{
This function predicts the optimal treatments with model of psi-Learning,which is estimated from \code{\link{psiITR}} or \code{\link{psiITR_VS}}
}
\examples{
          n=100;p=5;ntest=200
          X=replicate(p,runif(n, min = -1, max = 1))
          A=2*rbinom(n, 1, 0.5)-1
          T=cbind(rep(1,n,1),X)\%*\%c(1,2,1,0.5,rep(0,1,p-3))
          T0=(cbind(rep(1,n,1),X)\%*\%c(0.54,-1.8,-1.8,rep(0,1,p-2)))*A
          R=as.vector(rnorm(n,mean=0,sd=1)+T+T0)
          cv_psi_Linear<-cv.psiITR(X,A,R,m=5,kernel='linear',kappa.ratio=0.01,kappa.max=1.5,nkappa=10,tau=0.1,maxit=100, tol=1e-5)
          cv_psi_Rbf<-cv.psiITR(X,A,R,m=5,kernel='rbf',kappa.ratio=0.01,kappa.max=1,nkappa=10,tau=0.1,maxit=100, tol=1e-5)
          Xtest=replicate(p,runif(ntest, min = -1, max = 1))
          pre.psiLin=pre.psi(cv_psi_Linear,Xtest)
          sig=Sig_est(X,A)
          pre.psiRbf=pre.psi(cv_psi_Rbf,Xtest,X,kernel='rbf',sigma=sig)
}
