% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialEst.R
\name{psi_Init}
\alias{psi_Init}
\title{Initial value estimate for the psiLearn.}
\usage{
psi_Init(X,A,R,tau=0.1,kappa=seq(0.01,1,length.out=10),sigma=NULL,kernel='linear',lambda=NULL,VS=FALSE)
}
\arguments{
\item{X}{\eqn{n} by \eqn{p} input matrix.}

\item{A}{a vector of n entries coded 1 and -1 for the treatment assignments.}

\item{R}{a vector of outcome variable, larger is more desirable.}

\item{tau}{tuning parameter for the loss function in psi-Learn}

\item{kappa}{tunning parameter to control the complexity of the decision function, a seq of kappa is set as the default.}

\item{sigma}{bandwidth parameter when the kernel is 'rbf',if not provided,can be estimated from \code{\link{Sig_est}}}

\item{kernel}{kernel function for pai-Learn, can be 'linear' or 'rbf' (radial basis kernel), default is 'linear'.
When 'rbf' is specified, one can specify the sigma parameter for radial basis kernel or estimate from \code{\link{Sig_est}}.}

\item{lambda}{when using the variable selection method for inital estimate, lambda is regarded as the tunning parameter in controlling the complexity in L1 penalty}

\item{VS}{wheter to use the variable selection method as the inital estimate,default is FALSE}
}
\value{
It returns initial estimated coefficients for psiLearn with the best tuning parameters picked by cross validation.
 \item{w}{the coefficent for the decision function.}
 \item{bias}{the intercept in both the linear case and the kernel case.}
}
\description{
Initial value estimate for both psi-linear,psi-kernel,and psi-linear-VS.
}
\seealso{
\code{\link{tlp_S}}
}
\author{
MingyangLiu <liux3941@umn.edu>
}
