% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialEst.R
\name{resEst}
\alias{resEst}
\title{Residual estimate after subtracting the linear main effect}
\usage{
resEst(X,R,p)
}
\arguments{
\item{X}{\eqn{n} by \eqn{p} input matrix.}

\item{R}{a vector of outcome variable, larger is more desirable.}

\item{p}{a vector of probability of P(A|X) for each observation.}
}
\value{
It returns the residual in psi-Learn
 \item{res}{the residual in psi-Learn}
}
\description{
calculate the residual  after subtracting the linear main effect
}
\author{
MingyangLiu <liux3941@umn.edu>
}
