% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialEst.R
\name{Sig_est}
\alias{Sig_est}
\title{inverse bandwidth parameter estimate for the rbf kernel in psiLearn}
\usage{
Sig_est(X,A)
}
\arguments{
\item{X}{n*p input matrix.}

\item{A}{a vector of n entries coded 1 and -1 for the treatment assignments.}
}
\value{
It returns  estimated bandwidth parameter sigma for rbf kernel when it is not provided by the user
 \item{sigma}{the estimated bandwidth parameter.}
}
\description{
The inverse kernel width used in Gaussian kernel
}
\author{
MingyangLiu <liux3941@umn.edu>
}
