% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InitialEst.R
\name{tlp_S}
\alias{tlp_S}
\title{Cost function in psi-Learn}
\usage{
tlp_S(X,A,R,wt,w,tau=0.1,kappa=0.1,kernel='linear')
}
\arguments{
\item{X}{\eqn{n} by \eqn{p} input matrix.}

\item{A}{a vector of n entries coded 1 and -1 for the treatment assignments.}

\item{R}{a vector of outcome variable, larger is more desirable.}

\item{wt}{a vector of weights for each observation.}

\item{w}{coefficients for the decision function, the first element is the bias}

\item{tau}{tuning parameter for the loss function in psi-Learn}

\item{kappa}{tunning parameter to control the complexity of the decision function.}

\item{kernel}{kernel function for pai-Learn, can be 'linear' or 'rbf' (radial basis kernel), default is 'linear'.}
}
\value{
It returns the cost value in psi-Learn
 \item{cost}{the cost value in psi-Learn}
}
\description{
calculate the cost in psi-Learn with ridge penalty
}
\author{
MingyangLiu <liux3941@umn.edu>
}
